상세보기 페이지 제작완료 기능 구현 중
