+++ wherego/src/main/resources/mapper/accommodation.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.tencoding.wherego.repository.interfaces.AccommodationRepository">
+<mapper namespace="com.tencoding.wherego.repository.interfaces.accommodation.AccommodationRepository">
 	
-	<select id="findListById" resultType="com.tencoding.wherego.dto.AccommodationDto">
+	<select id="findListById" resultType="com.tencoding.wherego.dto.accommodation.AccommodationDto">
 		SELECT
 		ACC.room_no, ACC.room_name, ACC.room_view,
 		ACC.room_size, ACC.room_type, ACC.room_detail, 
@@ -13,6 +13,15 @@
 		LEFT JOIN accommodation_img AS IMG
 		ON ACC.room_no = IMG.room_no
 		GROUP BY room_no
+	</select>
+	 
+	<select id="findImgListById">
+		SELECT
+		IMG.img_url
+		FROM accommodation AS ACC
+		LEFT JOIN accommodation_img AS IMG
+		ON ACC.room_no = IMG.room_no
+        WHERE ACC.room_no = #{roomNo}
 	</select>
 	 
 </mapper>
