+++ wherego/src/main/webapp/WEB-INF/view/accommodation/acc_list.jsp
@@ -83,7 +83,7 @@
             </div>
             <c:forEach var="acc" items="${accList}">
 	            <div class="acc_list_contents">
-	                <div class="acc_list_contain">
+	                <div class="acc_list_contain" data-toggle="modal" data-target="#staticBackdrop" onclick="getImgList()">
 	                    <div class="acc_image">
 	                        <img src="${acc.imgUrl }" class="img-thumbnail" alt="...">
 	                    </div>
@@ -95,33 +95,106 @@
 	                    </div>
 	                </div>
 	                <div class="acc_list_button">
-	                    <button type="button" class="btn price_btn"  data-bs-toggle="modal" data-bs-target="#staticBackdrop">${acc.roomPrice } KRW</button>
+	                    <button type="button" class="price_btn">${acc.roomPrice } KRW</button>
 	                </div>
 	            </div>
             </c:forEach>
         </div>
     </div>
+    
     <!-- Modal -->
-	<div class="modal fade" id="staticBackdrop" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
-	  <div class="modal-dialog">
-	    <div class="modal-content">
-	      <div class="modal-header">
-	        <h5 class="modal-title" id="staticBackdropLabel">Modal title</h5>
-	        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-	      </div>
-	      <div class="modal-body">
-	        ...
-	      </div>
-	      <div class="modal-footer">
-	        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-	        <button type="button" class="btn btn-primary">Understood</button>
-	      </div>
-	    </div>
-	  </div>
-	</div>
+    <div class="modal fade" id="staticBackdrop" data-backdrop="static" data-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-lg modal-dialog-scrollable">
+        <div class="modal-content">
+            <div class="modal-header" style="background-color: #F5F5F5;">
+                <h5 style="font-weight: 600">디럭스 트윈</h5>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body" style="background-color: #F5F5F5;">
+                <div id="carouselExampleFade" class="carousel slide carousel-fade" data-ride="carousel">
+                    <div class="carousel-inner">
+                      <div class="carousel-item active">
+                          <img src="./img/jr_suite_family_twin_1.jpg" class="d-block w-100" alt="">
+                      </div>
+                      <div class="carousel-item">
+                          <img src="./img/jr_suite_family_twin_2.jpg" class="d-block w-100" alt="">
+                      </div>
+                      <div class="carousel-item">
+                          <img src="./img/jr_suite_family_twin_3.jpg" class="d-block w-100" alt="">
+                      </div>
+                      <div class="carousel-item">
+                          <img src="./img/jr_suite_family_twin_4.jpg" class="d-block w-100" alt="">
+                      </div>
+                      <div class="carousel-item">
+                          <img src="./img/jr_suite_family_twin_5.jpg" class="d-block w-100" alt="">
+                      </div>
+                    </div>
+                    <button class="carousel-control-prev" type="button" data-target="#carouselExampleFade" data-slide="prev">
+                      <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+                      <span class="sr-only">Previous</span>
+                    </button>
+                    <button class="carousel-control-next" type="button" data-target="#carouselExampleFade" data-slide="next">
+                      <span class="carousel-control-next-icon" aria-hidden="true"></span>
+                      <span class="sr-only">Next</span>
+                    </button>
+                </div>
+                <div class="acc_detail_div">
+                    <h3>디럭스 트윈</h3>
+                    <div>여유로운 휴식 공간이 돋보이는 롯데호텔 부산 디럭스 룸은 끊임 없는 리뉴얼로 최상의 컨디션을 유지하고 있습니다. 객실 내에 마련된 대형 데스크는 출장지에서도 빈틈 없는 업무 처리를 돕습니다. 창 밖으로 부산 한 중심의 화려한 시티뷰를 감상하실 수 있는 롯데호텔 부산 디럭스 룸에서 완벽한 여행을 꾸며 보세요.</div>
+                </div>
+                <div class="acc_detail_exp">
+                    <div class="acc_detail_info">
+                        <div class="acc_detail_info_title">객실 정보</div>
+                        <div>
+                            <div class="acc_detail_info_content">시티뷰</div>
+                            <div class="acc_detail_info_content">33.68~36.90 ㎡</div>
+                        </div>
+                    </div>
+                    <div class="br_div"></div>
+                    <div class="acc_detail_info">
+                        <div class="acc_detail_info_title">어메니티</div>
+                        <div>
+                            <div class="acc_detail_info_content">
+                                <div class="acc_detail_info_content_title">일반</div>
+                                <div>32'' LCD TV, 손전등, 슬리퍼, 구둣주걱, 구두클리너, 전화기, 티포트, 금고, 냉장고, 대형옷장, 미니바, 소화기</div>
+                            </div>
+                            <div class="acc_detail_info_content">
+                                <div class="acc_detail_info_content_title">욕실</div>
+                                <div>스폰지, 목욕 가운, 비데, 헤어드라이어, 욕조, 욕실용품, 면봉</div>
+                            </div>
+                            <div class="acc_detail_info_content">
+                                <div class="acc_detail_info_content_title">기타</div>
+                                <div>무료 생수 1일 2병, 보이스 메일 서비스, 무료 차(TEA) 서비스</div>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="br_div"></div>
+                    <div class="acc_detail_info">
+                        <div class="acc_detail_info_title">추가정보</div>
+                        <div class="acc_detail_info_content">
+                            <div class="acc_detail_info_content">전 객실 금연층 운영</div>
+                            <div class="acc_detail_info_content">조식 가격은 1인 ￦57,000입니다. (세금 및 봉사료 포함, 미이용 시에도 환불 불가)</div>
+                            <div class="acc_detail_info_content">보조 침대(Extra-Bed) 요청 시 ￦50,000이 가산됩니다.</div>
+                            <div class="acc_detail_info_content">모든 요금에는 세금 10%, 봉사료 10%(합계 21%)가 부과됩니다.</div>
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-exclamation-circle" viewBox="0 0 16 16">
+                                <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z"/>
+                                <path d="M7.002 11a1 1 0 1 1 2 0 1 1 0 0 1-2 0zM7.1 4.995a.905.905 0 1 1 1.8 0l-.35 3.507a.552.552 0 0 1-1.1 0L7.1 4.995z"/>
+                            </svg>
+                            &nbsp객실 내 구성은 객실 층 및 호텔 사정에 따라 사진과 다를 수 있습니다.
+                        </div>
+                    </div>
+                </div>
+                
+            </div>
+        </div>
       
     <script type="text/javascript">
-        
+        function getImgList() {
+        	let roomNumber = document.getElementById('roomNumber').value;
+            console.log(roomNumber);
+        }
     </script>
 
   </body>
